code/requirements.txt
sagemaker==2.16.3
tensorflow-gpu==2.0
tensorflow-gpu-estimator==2.1.0
tensorflow-estimator==2.1.0
tensorflow-hub==0.10.0
bert-for-tf2==0.14.7
bert-tensorflow==1.0.4
pydot==1.4.1
graphviz==0.15
boto3==1.16.17
scikit-learn==0.22
pandas==1.0.1